Move styles into styling file
